Remove 'recurse' parameter from Inline.
The way it is implemented currently is bad because while we're inlining
to a graph G, we are also mutating all the graphs that are being
inlined. The problem is that the graphs we're inlining are usually the
original graphs of functions, so we're silently changing them behind the
scenes, and we don't have a way to recover 'unoptimized' graphs
afterwards.

ghstack-source-id: 03696ec310380109fb420bb1e87c297c1cbacc4e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/26487